split utils to utils and process
