+++ cimr/processor/__init__.py
@@ -1,6 +1,6 @@
 
-from .utils import Infiler
-from .utils import Integrator
+from .process import Infiler
+from .process import Integrator
 from .query import Querier
 from .tad import Tadpole
 from .yamler import Yamler
